% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigDataPE.R
\name{bdpe_list_tokens}
\alias{bdpe_list_tokens}
\title{List all datasets with stored tokens}
\usage{
bdpe_list_tokens()
}
\value{
A vector of strings containing the names of datasets.
}
\description{
This function returns a list of datasets with securely stored
authentication tokens.
}
\examples{
bdpe_list_tokens()
}
