% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigDataPE.R
\name{bdpe_get_token}
\alias{bdpe_get_token}
\title{Retrieve the token associated with a specific dataset}
\usage{
bdpe_get_token(base_name)
}
\arguments{
\item{base_name}{The name of the dataset.}
}
\value{
A string containing the authentication token, or \code{NULL} if the token is not found.
}
\description{
This function retrieves the authentication token securely
stored for a specific dataset. If the token is not found,
it returns \code{NULL} and prints a message instead of throwing an error.
}
\examples{
token <- bdpe_get_token("education_dataset")
}
