% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigDataPE.R
\name{bdpe_list_tokens}
\alias{bdpe_list_tokens}
\title{List all datasets that have stored tokens in environment variables}
\usage{
bdpe_list_tokens()
}
\value{
A character vector of dataset names with stored tokens.
If no tokens are found, an empty vector is returned and
a message is printed.
}
\description{
This function returns a character vector of dataset names
that have their tokens stored in environment variables.
Specifically, it looks for variables that begin with the prefix
\code{"BigDataPE_"}.
}
\examples{
bdpe_list_tokens()

}
