% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigDataPE.R
\name{bdpe_remove_token}
\alias{bdpe_remove_token}
\title{Remove the token associated with a specific dataset}
\usage{
bdpe_remove_token(base_name)
}
\arguments{
\item{base_name}{The name of the dataset.}
}
\value{
No return value. If the token is found, it is removed. If not, a message is displayed.
}
\description{
This function removes the securely stored authentication
token for a specific dataset. If the token is not found,
it prints a message and does not throw an error.
}
\examples{
bdpe_remove_token("education_dataset")
}
