% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigDataPE.R
\name{bdpe_store_token}
\alias{bdpe_store_token}
\title{Store a token securely for a specific dataset}
\usage{
bdpe_store_token(base_name, token)
}
\arguments{
\item{base_name}{The name of the dataset.}

\item{token}{The authentication token for the dataset.}
}
\value{
No return value. The token is securely stored.
}
\description{
This function securely stores an authentication token
for a specific dataset using the \code{keyring} package.
}
\examples{
bdpe_store_token("education_dataset", "your-token-here")
}
