% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigDataPE.R
\name{parse_queries}
\alias{parse_queries}
\title{Constructs a URL with query parameters}
\usage{
parse_queries(url, query_list)
}
\arguments{
\item{url}{The base URL to which query parameters will be added.}

\item{query_list}{A list of query parameters to be added to the URL.}
}
\value{
The complete URL with the query parameters appended.
}
\description{
This function appends a list of query parameters to a base URL.
}
\examples{
parse_queries("https://www.example.com", list(param1 = "value1", param2 = "value2"))
}
